Move CCTexture2D & CCSpriteFrame to assets
