+++ cocos2d/core/assets/CCSpriteFrame.js
@@ -26,6 +26,7 @@
  ****************************************************************************/
 
 var EventTarget = require("../event/event-target");
+var textureUtil = require('../utils/texture-util');
 
 /**
  * !#en
@@ -396,7 +397,7 @@ var SpriteFrame = cc.Class(/** @lends cc.SpriteFrame# */{
 
     _loadTexture: function () {
         if (this._textureFilename) {
-            var texture = cc.textureUtil.loadImage(this._textureFilename);
+            var texture = textureUtil.loadImage(this._textureFilename);
             this._refreshTexture(texture);
         }
     },
@@ -424,7 +425,7 @@ var SpriteFrame = cc.Class(/** @lends cc.SpriteFrame# */{
             if (!this._texture.loaded) {
                 // load exists texture
                 this._refreshTexture(this._texture);
-                cc.textureUtil.postLoadTexture(this._texture);
+                textureUtil.postLoadTexture(this._texture);
             }
         }
         else if (this._textureFilename) {
