+++ cocos2d/core/assets/CCTexture2D.js
