+++ cocos2d/core/assets/index.js
@@ -31,10 +31,9 @@ require('./CCPrefab');
 require('./CCAudioClip');
 require('./CCScripts');
 require('./CCSceneAsset');
-require('../sprites/CCSpriteFrame');
-require('../textures/CCTexture2D');
+require('./CCSpriteFrame');
+require('./CCTexture2D');
 require('./CCTTFFont');
 require('./CCSpriteAtlas');
 require('./CCBitmapFont');
 require('./CCLabelAtlas');
-//require('./CCTextAsset');
