+++ gulp/tasks/engine.js
@@ -40,8 +40,6 @@ const Optimizejs = require('gulp-optimize-js');
 
 var jsbSkipModules = [
     '../../cocos2d/core/CCGame',
-    // '../../cocos2d/core/textures/CCTexture2D',
-    // '../../cocos2d/core/sprites/CCSpriteFrame',
     '../../cocos2d/core/load-pipeline/audio-downloader',
     '../../cocos2d/core/platform/CCInputManager.js',
     '../../cocos2d/core/platform/CCVisibleRect.js',
